Move subgroup selection definition into own header file.
